+++ CHANGELOG.md
@@ -1,10 +1,32 @@
 # ChangeLog for Light Coding
 
+## [1.2 Version] - 2026-06-09
+## Added
+- Log Added
+- New Exploit Functions: BlockRE, UnBlockRE, getrawmetatable, getfenv, fireclickdetector, firetouchinterest, fireproximitypromtp, firesignal, setfflag, getfflag, gethwid, clonoref, replicatesignal, gethiddenproperty, sethiddenproperty, gethui
+- New Method: LightCoding:AddCustomFunc()
+- New Method: LightCoding:AddCustomMethod()
+- New Method: LightCoding:ClearLog()
+- New Method: LightCoding:CallFunc()
+- New Method: LightCoding:Delete()
+- New Method: LightCoding:GetFunc()
+- New Variable: LightCoding.Settings.Loaded or getgenv().LCLoaded or _G.LCLoaded
+
+## Changed
+
+- Now you can LightCoding:CallFunc("getgenv", "hi") no longer LightCoding.Functs.getgenv.hi
+- No longer restart script with getgenv().LCDebug = true to enable Debug
+
+## Removed
+
+ - Remove err function
+
+---
+
 ## [1.1 Version] - 2026-05-16
 ### Added
 
 - New Exploit Functions: getgenv, executor, clipboard, writefile, readfile, isfile, makefolder, isfolder, hookfunction, hookmetamethod, getnamecallmethod, checkaller, newcclosure, readonly, setreadonly, makewriteable, isreadonly, getgc
-- New Debug Function: console
 - Added Debug
 - New Sections Info, Settings
 - Added new Automatic Debug
@@ -23,3 +45,4 @@
 - First publication of the repository
 - Basic Functions: say, warnmsg, err, service, create
 - Exploit Functions: pfireclickd, pfiretouchinterest, pfireproximitypromt
+-